Fix: merge outbound warehouse selection tests into outbound_flow spec
