+++ apps/e2e/integration/note-transactions/outbound_note_transactions.spec.ts
@@ -991,85 +991,3 @@ test("warehouse dropdown should not show options for deleted warehouses", async 
 	await expect(dropdown.locator("div", { hasText: "Warehouse 1" })).not.toBeVisible();
 	await expect(dropdown.locator("div", { hasText: "Warehouse 2" })).not.toBeVisible();
 });
-
-test("ForceWithdrawal dialog should not show warehouses with available stock", async ({ page }) => {
-	// Setup: Create warehouses with different stock levels
-	const dbHandle = await getDbHandle(page);
-	const isbn = "1234567890";
-
-	// Warehouse 1 has 2 copies
-	await dbHandle.evaluate(upsertWarehouse, { id: 1, displayName: "Warehouse 1" });
-	await dbHandle.evaluate(createInboundNote, { id: 2, warehouseId: 1 });
-	await dbHandle.evaluate(addVolumesToNote, [2, { isbn, quantity: 2, warehouseId: 1 }] as const);
-	await dbHandle.evaluate(commitNote, 2);
-
-	// Warehouse 2 has 0 copies (no stock)
-	await dbHandle.evaluate(upsertWarehouse, { id: 2, displayName: "Warehouse 2" });
-
-	// Warehouse 3 has 1 copy
-	await dbHandle.evaluate(upsertWarehouse, { id: 3, displayName: "Warehouse 3" });
-	await dbHandle.evaluate(createInboundNote, { id: 3, warehouseId: 3 });
-	await dbHandle.evaluate(addVolumesToNote, [3, { isbn, quantity: 1, warehouseId: 3 }] as const);
-	await dbHandle.evaluate(commitNote, 3);
-
-	// Add a transaction for the book
-	const content = getDashboard(page).content();
-	const scanField = content.scanField();
-	const entries = content.table("outbound-note");
-
-	await scanField.add(isbn);
-
-	// The transaction should be assigned to a warehouse with stock
-	await entries.assertRows([{ isbn, quantity: 1, warehouseName: "Warehouse 1" }]);
-
-	// Add another transaction to exceed available stock in Warehouse 1
-	await scanField.add(isbn);
-	await scanField.add(isbn);
-
-	// Now we should have one transaction with Warehouse 1 (2 copies) and one with Warehouse 3 (1 copy)
-	await entries.assertRows([
-		{ isbn, quantity: 1, warehouseName: "Warehouse 3" },
-		{ isbn, quantity: 2, warehouseName: "Warehouse 1" }
-	]);
-
-	// Add one more to exceed all available stock
-	await scanField.add(isbn);
-
-	// Now we should have one transaction with no warehouse assigned
-	await entries.assertRows([
-		{ isbn, quantity: 1, warehouseName: "" },
-		{ isbn, quantity: 1, warehouseName: "Warehouse 3" },
-		{ isbn, quantity: 2, warehouseName: "Warehouse 1" }
-	]);
-
-	// Open the warehouse selector dropdown for the unassigned transaction
-	await entries.row(0).field("warehouseName").click();
-
-	// Click the "Force Withdrawal" button
-	const dropdown = page.getByTestId("dropdown-menu");
-	await dropdown.locator("button", { hasText: "Force withdrawal" }).click();
-
-	// A dialog should appear with warehouse options
-	const forceWithdrawalDialog = page.getByRole("dialog");
-	const warehouseSelect = forceWithdrawalDialog.locator("#warehouse-force-withdrawal");
-
-	// Check that Warehouse 2 (no stock) is available in the dropdown
-	await expect(warehouseSelect.locator('option[value="2"]')).toBeVisible();
-
-	// Check that Warehouse 1 and 3 (with stock) are NOT available in the dropdown
-	// They should be hidden with the "hidden" class
-	await expect(warehouseSelect.locator('option[value="1"]')).toHaveClass(/hidden/);
-	await expect(warehouseSelect.locator('option[value="3"]')).toHaveClass(/hidden/);
-
-	// Select Warehouse 2 and confirm
-	await warehouseSelect.selectOption({ label: "Warehouse 2" });
-	await forceWithdrawalDialog.getByRole("button", { name: "Confirm" }).click();
-
-	// Verify the transaction is now assigned to Warehouse 2 as a forced withdrawal
-	await entries.assertRows([
-		{ isbn, quantity: 1, warehouseName: "Warehouse 2" },
-		{ isbn, quantity: 1, warehouseName: "Warehouse 3" },
-		{ isbn, quantity: 2, warehouseName: "Warehouse 1" }
-	]);
-	await expect(content.table("outbound-note").row(0).field("warehouseName")).toContainText("Forced");
-});
